Move DDO implementation to its own module
