+++ aries_vcx/src/messages/connection/mod.rs
@@ -1,5 +1,4 @@
 pub mod did;
-pub mod did_doc;
 pub mod invite;
 pub mod problem_report;
 pub mod request;
