+++ aries_vcx/src/protocols/connection/invitee/states/responded.rs
@@ -1,4 +1,4 @@
-use crate::messages::connection::did_doc::DidDoc;
+use crate::did_doc::DidDoc;
 use crate::messages::connection::problem_report::ProblemReport;
 use crate::messages::connection::request::Request;
 use crate::messages::connection::response::SignedResponse;
