+++ aries_vcx/src/protocols/connection/inviter/states/responded.rs
@@ -6,7 +6,7 @@ use indy_sys::WalletHandle;
 use crate::error::prelude::*;
 use crate::messages::a2a::A2AMessage;
 use crate::messages::ack::Ack;
-use crate::messages::connection::did_doc::DidDoc;
+use crate::did_doc::DidDoc;
 use crate::messages::connection::problem_report::ProblemReport;
 use crate::messages::connection::response::SignedResponse;
 use crate::messages::trust_ping::ping::Ping;
