+++ aries_vcx/src/protocols/connection/util.rs
@@ -6,7 +6,7 @@ use indy_sys::WalletHandle;
 use crate::error::prelude::*;
 use crate::global::settings;
 use crate::messages::a2a::A2AMessage;
-use crate::messages::connection::did_doc::DidDoc;
+use crate::did_doc::DidDoc;
 use crate::messages::trust_ping::ping::Ping;
 use crate::messages::trust_ping::ping_response::PingResponse;
