+++ aries_vcx/src/utils/encryption_envelope.rs
@@ -7,7 +7,7 @@ use crate::error::prelude::*;
 use crate::global::settings;
 use crate::libindy::utils::crypto;
 use crate::messages::a2a::A2AMessage;
-use crate::messages::connection::did_doc::DidDoc;
+use crate::did_doc::DidDoc;
 use crate::messages::forward::Forward;
 
 #[derive(Debug)]
@@ -142,7 +142,7 @@ pub mod unit_tests {
     use crate::libindy::utils::crypto::create_key;
     use crate::libindy::utils::test_setup::create_trustee_key;
     use crate::messages::ack::test_utils::_ack;
-    use crate::messages::connection::did_doc::test_utils::*;
+    use crate::did_doc::test_utils::*;
     use crate::utils::devsetup::SetupEmpty;
 
     use super::*;
