+++ aries_vcx/src/utils/mod.rs
@@ -4,7 +4,7 @@ use indy_sys::WalletHandle;
 
 use crate::error::VcxResult;
 use crate::messages::a2a::A2AMessage;
-use crate::messages::connection::did_doc::DidDoc;
+use crate::did_doc::DidDoc;
 use crate::utils::encryption_envelope::EncryptionEnvelope;
 
 #[macro_use]
